refactor: product 정보를 DetailProduct 컴포넌트로 분리
